% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{phyloseq_to_biom}
\alias{phyloseq_to_biom}
\title{Check status of the project
Load folder of microbiome data. The folder must have asv.tab, taxonomy.tsv, metadata.tsv files}
\usage{
phyloseq_to_biom(physeq)
}
\value{
phyloseq object
}
\description{
Check status of the project
Load folder of microbiome data. The folder must have asv.tab, taxonomy.tsv, metadata.tsv files
}
