% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{Project}
\alias{Project}
\title{Project Class}
\description{
An R6 class to handle project data operations.
To use start with my_project <- ProjectA$new("path/to/your/folder")
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{folder_path}}{The path to the project folder}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Project-new}{\code{Project$new()}}
\item \href{#method-Project-validate}{\code{Project$validate()}}
\item \href{#method-Project-get_alpha_data}{\code{Project$get_alpha_data()}}
\item \href{#method-Project-get_beta_data}{\code{Project$get_beta_data()}}
\item \href{#method-Project-get_taxa_data}{\code{Project$get_taxa_data()}}
\item \href{#method-Project-get_metadata}{\code{Project$get_metadata()}}
\item \href{#method-Project-get_filtered_metadata}{\code{Project$get_filtered_metadata()}}
\item \href{#method-Project-get_ecfunc_data}{\code{Project$get_ecfunc_data()}}
\item \href{#method-Project-get_kofunc_data}{\code{Project$get_kofunc_data()}}
\item \href{#method-Project-get_ko_sankey}{\code{Project$get_ko_sankey()}}
\item \href{#method-Project-get_ptfunc_data}{\code{Project$get_ptfunc_data()}}
\item \href{#method-Project-get_project_overview}{\code{Project$get_project_overview()}}
\item \href{#method-Project-get_project_id}{\code{Project$get_project_id()}}
\item \href{#method-Project-clone}{\code{Project$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-new"></a>}}
\if{latex}{\out{\hypertarget{method-Project-new}{}}}
\subsection{Method \code{new()}}{
Create a new ProjectA object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$new(folder_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{folder_path}}{The path to the project folder}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Project-validate}{}}}
\subsection{Method \code{validate()}}{
Validate the presence of required files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-get_alpha_data"></a>}}
\if{latex}{\out{\hypertarget{method-Project-get_alpha_data}{}}}
\subsection{Method \code{get_alpha_data()}}{
Get alpha diversity data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_alpha_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list containing alpha diversity and metadata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-get_beta_data"></a>}}
\if{latex}{\out{\hypertarget{method-Project-get_beta_data}{}}}
\subsection{Method \code{get_beta_data()}}{
Get beta diversity data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_beta_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list containing PCoA, NMDS, and metadata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-get_taxa_data"></a>}}
\if{latex}{\out{\hypertarget{method-Project-get_taxa_data}{}}}
\subsection{Method \code{get_taxa_data()}}{
Get taxonomic data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_taxa_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list containing ASV, taxonomy, and metadata
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-Project-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_metadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-get_filtered_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-Project-get_filtered_metadata}{}}}
\subsection{Method \code{get_filtered_metadata()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_filtered_metadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-get_ecfunc_data"></a>}}
\if{latex}{\out{\hypertarget{method-Project-get_ecfunc_data}{}}}
\subsection{Method \code{get_ecfunc_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_ecfunc_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-get_kofunc_data"></a>}}
\if{latex}{\out{\hypertarget{method-Project-get_kofunc_data}{}}}
\subsection{Method \code{get_kofunc_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_kofunc_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-get_ko_sankey"></a>}}
\if{latex}{\out{\hypertarget{method-Project-get_ko_sankey}{}}}
\subsection{Method \code{get_ko_sankey()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_ko_sankey()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-get_ptfunc_data"></a>}}
\if{latex}{\out{\hypertarget{method-Project-get_ptfunc_data}{}}}
\subsection{Method \code{get_ptfunc_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_ptfunc_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-get_project_overview"></a>}}
\if{latex}{\out{\hypertarget{method-Project-get_project_overview}{}}}
\subsection{Method \code{get_project_overview()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_project_overview()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-get_project_id"></a>}}
\if{latex}{\out{\hypertarget{method-Project-get_project_id}{}}}
\subsection{Method \code{get_project_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$get_project_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Project-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
